requirements.txt
click==8.0.4
cycler==0.11.0
Flask==2.0.3
fonttools==4.29.1
itsdangerous==2.1.0
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.1.0
matplotlib==3.5.1
numpy==1.21.5
packaging==21.3
Pillow==9.0.1
pyparsing==3.0.7
python-dateutil==2.8.2
six==1.16.0
torch==1.10.2
torchvision==0.11.3
typing-extensions==4.1.1
Werkzeug==2.0.3